+++ example/lib/key/key_model.dart
@@ -2,5 +2,6 @@ import 'package:flutter/widgets.dart';
 import 'package:mvc/base/base_model.dart';
 
 class KeyModel extends BaseModel {
+  GlobalKey scrollKey = GlobalKey();
   GlobalKey key = GlobalKey();
 }