+++ lib/base/base_page.dart
@@ -33,12 +33,12 @@ abstract class BasePage<T extends BaseController, M extends BaseModel>
   String get tagSymbol => "page";
 
   @override
-  String? get tagHotReload => null;
+  String? get tagRepeat => null;
 
   @override
   String? get tag {
     if (!reuseController) {
-      if (tagHotReload != null) return tagHotReload;
+      if (tagRepeat != null) return tagRepeat;
       return PageStack.current(tagSymbol);
     }
     return null;
@@ -109,7 +109,7 @@ abstract class BasePage<T extends BaseController, M extends BaseModel>
   /// 1.页面需要重复跳转 且reuseController为false
   /// 2.页面中引用了controller或者model中声明且已经创建的GlobalKey对象
   @protected
-  void initHotReloadTag(String? tag) {}
+  void initRepeatTag(String? tag) {}
 
   /// 仅在 createState 为 TabControllerState 时使用
   @protected
@@ -149,7 +149,7 @@ class AutoDisposeState<P extends BasePage, T extends BaseController>
   void initState() {
     if (!widget.reuseController) {
       currentTag = PageStack.push(widget.tagSymbol);
-      widget.initHotReloadTag(currentTag);
+      widget.initRepeatTag(currentTag);
     }
     Get.put<T>(widget.binding as T,
         tag: currentTag, permanent: widget.permanentController);
@@ -166,7 +166,7 @@ class AutoDisposeState<P extends BasePage, T extends BaseController>
 
   @override
   Widget build(BuildContext context) {
-    widget.initHotReloadTag(currentTag);
+    widget.initRepeatTag(currentTag);
     return widget.build(context);
   }